feat: 내가 좋아요 했는지 표시기능 권한 설정

+++ backend/rush/src/main/java/rush/rush/dto/ArticleResponse.java
@@ -18,5 +18,5 @@ public class ArticleResponse {
     private Double longitude;
     private AuthorResponse author;
     private Timestamp createDate;
-    private int totalLikes;
+    private Long totalLikes;
 }
